Removed environment files, added new packages and refactored the project structure
This commit contains several enhancements to the project. Firstly, the miscellaneous environment-related files `misc.xml` and `zhivaya_bot.iml` have been removed. This is because they were not critical to the function of the project, and their removal simplifies the repository and ensures it remains clean and focused.

Secondly, the `Dockerfile` has been created to contain the necessary instructions